Rename "SliderButtons" to "BannerArrows" for more clarity.
