+++ frontend/src/components/BannerArrows/NextArrow.tsx
