+++ frontend/src/components/BannerArrows/PrevArrow.tsx
