+++ frontend/src/components/BannerArrows/SliderButton.scss
